requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django
djangorestframework
envconfiguration
fontawesomefree
requests
xlsxwriter
mysqlclient
pyyaml
celery
redis
djangorestframework-simplejwt
python-docx
ntplib
reportlab